+++ include/core/distanceMetrics/DistanceMetric.hpp
@@ -39,8 +39,8 @@ template <unsigned int dim>
 class DistanceMetric : public Identifier {
   public:
     DistanceMetric(const std::string &name = "L2 Metric");
-    float calcDist(const std::shared_ptr<Node<dim>> &source, const std::shared_ptr<Node<dim>> &target) const;
-    virtual float calcDist(const Vector<dim> &source, const Vector<dim> &target) const = 0;
+    double calcDist(const std::shared_ptr<Node<dim>> &source, const std::shared_ptr<Node<dim>> &target) const;
+    virtual double calcDist(const Vector<dim> &source, const Vector<dim> &target) const = 0;
 };
 
 /*!
@@ -62,7 +62,7 @@ DistanceMetric<dim>::DistanceMetric(const std::string &name) : Identifier(name) 
 *  \date       2017-01-02
 */
 template <unsigned int dim>
-float DistanceMetric<dim>::calcDist(const std::shared_ptr<Node<dim>> &source, const std::shared_ptr<Node<dim>> &target) const {
+double DistanceMetric<dim>::calcDist(const std::shared_ptr<Node<dim>> &source, const std::shared_ptr<Node<dim>> &target) const {
     return calcDist(source->getValues(), target->getValues());
 }
